Make activities pagination consistent with other endpoints
In an effort to make our pagination consistent across our API endpoints, we found that this endpoint is using a different way of paginating. We're changing the parameter from `first` to `page[size]` and will be introducing links to next a potential next page in the response.

As Github has been the only user of this endpoint, we're updating this client to be compatible with the new pagination. All other endpoints will remain the same.